Added adjacency list representation of graph
This code is an implementation of an adjacency list representation of an undirected graph.
The addEdge method takes in the graph (as a list of lists), and the node indices u and v, and adds v to the list of u and u to the list of v. This represents an undirected edge between the two nodes.
The printGraph method iterates through each node in the graph and prints out the nodes it is connected to.