Linear Regression inC
In this implementation, the user is first prompted to enter the number of data points they will be providing. Then, the program reads in each x and y value provided by the user, storing them in arrays and updating the sum of x values, y values, their product, and their square at each iteration.